+++ src/common/hash-functions.hpp
@@ -13,17 +13,17 @@
 namespace std
 {
     template<>
-    struct hash<std::pair<size_t, size_t>> {
-        size_t operator()(const std::pair<size_t,size_t> key) const;
+    struct hash<std::pair<uint64_t, uint64_t>> {
+        uint64_t operator()(const std::pair<uint64_t,uint64_t> key) const;
     };
 
     template<>
-    struct hash<std::tuple<size_t, size_t>> {
-        size_t operator()(const std::tuple<size_t,size_t> key) const;
+    struct hash<std::tuple<uint64_t, uint64_t>> {
+        uint64_t operator()(const std::tuple<uint64_t,uint64_t> key) const;
     };
 
     template<>
-    struct hash<std::vector<size_t>> {
-        size_t operator()(const std::vector<size_t> vec) const;
+    struct hash<std::vector<uint64_t>> {
+        uint64_t operator()(const std::vector<uint64_t> vec) const;
     };
 }
